Update User service
