--- user-service/.dockerignore
@@ -1,3 +1,0 @@
-node_modules/
-npm-debug.log
-.DS_Store
