+++ user-service/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@grpc/grpc-js": "^1.12.2",
     "@grpc/proto-loader": "^0.7.13",
-    "@prisma/client": "^5.19.1",
+    "@prisma/client": "^5.21.1",
     "axios": "^1.7.7",
     "bcrypt": "^5.1.1",
     "jsonwebtoken": "^9.0.2",
