+++ user-service/src/server.ts
@@ -1,13 +1,14 @@
 import * as grpc from '@grpc/grpc-js';
 import * as protoLoader from '@grpc/proto-loader';
 import { PrismaClient } from '@prisma/client';
-import { getAuthenticatedUserId, createAuthCookie, verifyTicket } from './libs/auth';
+import { getAuthenticatedUserId, createAuthCookie } from './libs/token';
+import { verifyTicket } from './libs/auth';
 
 const prisma = new PrismaClient();
 
 console.log("Database connected");
 
-const PROTO_PATH = '../../proto/user.proto';
+const PROTO_PATH = '../proto/user.proto';
 
 const packageDefinition = protoLoader.loadSync(PROTO_PATH, {
   keepCase: true,
